requirements.txt
# Core extraction + RAG stack
openai>=1.52.0
instructor>=1.4.0
llama-index-core>=0.12.0
llama-index-vector-stores-chroma>=0.4.0
llama-index-embeddings-huggingface>=0.5.0

# Vector store and embeddings
sentence-transformers>=3.0.0
chromadb>=0.5.0
diskcache>=5.6.0

# Text processing
nltk>=3.8.1
textblob>=0.17.1
langdetect>=1.0.9
pdfplumber>=0.10.0
python-docx>=1.1.0

# Thematic analysis
bertopic>=0.16.0
keybert>=0.8.0

# Data processing and validation
pandas>=2.1.0
numpy>=1.24.0
pyyaml>=6.0.1
pydantic>=2.8.0

# Web interface
streamlit>=1.35.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
